Added DTO convertor for project and project controller

+++ src/main/java/todoapplication/service/impl/JpaProjectService.java
@@ -3,11 +3,13 @@ package todoapplication.service.impl;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import todoapplication.model.Project;
 import todoapplication.repository.ProjectRepository;
 import todoapplication.service.ProjectService;
 
+@Service
 public class JpaProjectService implements ProjectService{
 
 	@Autowired
@@ -29,13 +31,38 @@ public class JpaProjectService implements ProjectService{
 	}
 
 	@Override
-	public void delete(Long id) {
-		projectRepository.deleteById(id);
+	public Project delete(Long id) {
+		Project p = projectRepository.findOneById(id);
+		
+		if( p != null) {
+			
+			p.getProjectOwner().getProjects().remove(p);
+			
+			projectRepository.deleteById(id);
+			return p;
+		}
+		
+		return null;
 	}
 
 	@Override
 	public Project findByName(String name) {
 		return projectRepository.findByNameContainsIgnoreCase(name);
+	}
+
+	@Override
+	public Project findOneById(Long id) {
+		return projectRepository.findOneById(id);
+	}
+
+	@Override
+	public Project update(Project p) {
+		return projectRepository.save(p);
+	}
+
+	@Override
+	public List<Project> findByProjectOwner(Long id) {
+		return projectRepository.findByProjectOwner_Id(id);
 	}
 
 }
